move external_storage/util to tikv_util/stream
