+++ components/sst_importer/src/sst_importer.rs
@@ -26,8 +26,9 @@ use engine_traits::{
     EncryptionKeyManager, IngestExternalFileOptions, Iterator, KvEngine, SeekKey, SstReader,
     SstWriter, SstWriterBuilder, CF_DEFAULT, CF_WRITE,
 };
-use external_storage::{block_on_external_io, create_storage, url_of_backend, READ_BUF_SIZE};
+use external_storage::{create_storage, url_of_backend};
 use file_system::sync_dir;
+use tikv_util::stream::{block_on_external_io, READ_BUF_SIZE};
 use tikv_util::time::Limiter;
 use txn_types::{is_short_value, Key, TimeStamp, Write as KvWrite, WriteRef, WriteType};
