+++ components/tikv_util/Cargo.toml
@@ -22,6 +22,7 @@ prost-codec = [
 async-speed-limit = "0.3"
 backtrace = "0.3.9"
 byteorder = "1.2"
+bytes = "0.5.3"
 codec = { path = "../codec", default-features = false }
 configuration = { path = "../configuration" }
 chrono = "0.4"
@@ -32,6 +33,7 @@ derive_more = "0.99.3"
 fail = "0.4"
 fs2 = "0.4"
 futures = { version = "0.3", features = ["compat"] }
+futures-util = { version = "0.3", default-features = false, features = ["io"] }
 grpcio = { version = "0.7", default-features = false, features = ["openssl-vendored"] }
 lazy_static = "1.3"
 libc = "0.2"
@@ -42,6 +44,7 @@ openssl = "0.10"
 prometheus = { version = "0.10", features = ["nightly"] }
 quick-error = "1.2.3"
 rand = "0.7"
+rusoto_core = "0.45.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
